Refactor handles instantiation and callbacks
Now handles implement 'new' static method, so instead of

   UVAsync(loop)

we have

   UVAsync.new(loop)

This is faster, because we avoid creating a tuple for __init__, and
allows us to pass pointers to handles' constructors.

Handles that receive a callback (UVAsync, UVIdle, UVSignal, etc) now
receive it as a pointer to one-arg C function.