POC UVloop tracing, support for DNS traces
This is the first PR that implements POC that gives to uvloop the
capacity of emit traces that later can be collected by the upper layers.
What does this commit implement?

- DNS trace events triggered by the calls to the `getaddrinfo` loop
  method.
- The enabler and disabler context manager called `tracing`.
- The collector interface used to gather the trace events called
  `TracingCollector.

This POC assumes that the caller is the last responsible implementing
a context that isolates calling to traces that belong to different
tasks.

Missing pices in this PoC:

- Allow many collectors at a time.
- The exceptions have to be propagated at tracing level? perhaps, what
  should we do when there is an exception and we have to call to
`dns_request_end` ? right now a `None` is passed as argument