POC UVloop tracing, support for DNS traces and create Task traces
This is the first PR that implements POC that gives to uvloop the
capacity of emit traces that later can be collected by the upper layers.
What does this commit implement?

- DNS traces triggered by the calls to the `getaddrinfo` loop
  method.
- Create task traces emited by the `create_task` loop method.
- The enabler and disabler context manager called `tracing`.
- The collector interface used to gather the trace events called
  `TracingCollector.
- Capacity of install many nested collectors

This POC assumes that the caller is the last responsible implementing
a context that put together traces that belong to different tasks. Take
a look to this snippet [1]

Regarding the overhead of this POC, initial figures say that the performance
footprint while there are no collectors installed is almost negligible [2].

(master) Cost per task 5.2499999999417925e-06
(traces) No collectors: Cost task 5.240000000048895e-06
(traces) default method: Cost per task 5.3500000000349245e-06
(traces) Collector method: Cost per task 5.409999999974389e-06

There is still some pending stuff realted to this PR that is not solved:

- More robust tests for the DNS instrumentalization, right now params are not asserted.
- Instrumentalize `_getnameinfo`.
- More testing regarding `AddrInfoRequest`.

But before I would like to gather your toughts with this MR, I will make the proper comments in some parts of the code that can delicated.

[1] https://gist.github.com/pfreixes/5e9cccc4f5881de003243879edd5c470
[2] https://gist.github.com/pfreixes/18f0c678a9a018be0f376d4bef6d722a