Don't use non-existent UDPTransport._address attribute (fixes: #207)
If the sendto() method is called after closing the transport, it bombs
with an AttributeError as there is no `._address` attribute. Replacing
this by `.address` is not a solution either, as we need to exit in any
case otherwise the call to sendto() bombs again due to `.sock` having
been set to None.