setup.py: append environment CFLAGS to local CFLAGS variable
At the moment 'CFLAGS = ['-O2']' is passed to 'extra_compile_args' but
this way environment CFLAGS -O flag gets ignored. This way we can't
override -O CFLAGS for 'loop.pyx'. So let's append environment CFLAGS to
local CFLAGS variable that in order is assigned to 'extra_compile_args'.
This way we can override the default value of local variable CFLAGS.

Signed-off-by: Giulio Benetti <giulio.benetti@benettiengineering.com>