Update loop task factory typing with kwargs
For `PY311` the task factory will be called with the context kwarg. This
is not correctly reflected in the typing for `Loop.set_task_factory` and
`Loop.get_task_factory`. This commit addresses this with a protocol.
Since `uvloop` requires a minimum Python 3.8 version, in which
`typing.Protocol` is available; the more appropriate `typing.Unpack` for
kwargs is introduced in Python 3.11.