Compensate for asyncio changes in Python 3.13.3/3.14
In
https://github.com/python/cpython/commit/38a99568763604ccec5d5027f0658100ad76876f
(backported to 3.13 as
https://github.com/python/cpython/commit/7b0543ebe649aea11531e994289293f23f41064e),
Python moved the responsibility for setting the task name from
`asyncio.create_task` to the `Task` constructor.  While `uvloop` itself
is unaffected, this causes `test_set_task_name` to fail when run with
`asyncio`.  Compensate for that in this particular test.

It's possible that `uvloop`'s behaviour should be changed to match
`asyncio`'s depending on the Python version, but that seems like a more
involved change.  For now, this at least gets the tests passing again.

Most of this analysis and patch were from Martin Hostettler in
https://bugs.debian.org/1101258#24; I just tweaked the patch slightly to
ensure it still passes on older Python versions.